+++ src/fingerprint.ts
@@ -6,7 +6,7 @@ import { Fingerprint } from 'discord-verify';
  * Get the X-Fingerprint Header from Discord
  */
 const fingerprint = async (): Promise<Fingerprint> => {
-    const ContextProperties = Buffer.from(JSON.stringify({ location: 'Login' })).toString('base64')
+    const ContextProperties = Buffer.from(JSON.stringify({ location: 'Login' })).toString('base64');
     const res = await fetch('https://discordapp.com/api/v6/experiments', {
         headers: {
             'Accept': '*/*',