+++ src/util/delay.ts
@@ -1,0 +1,5 @@
+/**
+ * Sleep
+ * @param {number} ms Milliseconds to delay program by.
+ */
+export const delay = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));